refactor: clean-up printing via ToPrint typeclass (#1671)
This PR cleans up printing via the `ToPrint` typeclass

* Firstly, we rename the class to `DialectPrint` and generally rename a
lot of the methods with more informative names
* The class is also moved to the dedicated Print.lean file
* We add a lot more documentation
* We separate out a function which prints with an implicit
`builtin.module`, which is now called `Com.printModule` and a function
which prints normally, which is now called `Com.print`
* We remove a lot of unnecessary typeclass assumptions
* We refactor `Com.print` to return a `Format`, and use it internally to
ensure consistent indentation and alignment.
* We change the LLVM dialect print instance to no longer double print
the llvm. prefix, and fix the bug where flags were being printed as part
of the Op name

For now, we don't touch the other two printing methods, but in future
these will be unified, and the functions added in this PR will be the
canonical implementation