feat: DialectParse convenience class  (#1672)
This PR adds `DialectParse` as a more convenient way to define a parser
for a Dialect, and uses it to define a test dialect with multiple return
values.

The goal was primarily building the test dialect, but the ergonomics of
the `Transform` classes has been bothering me for a while, so I figured
I might as well fix that, too.