refactor: re-define Repr Com in terms of Com.print
This PR drops the old Repr instances for `Com` and `Expr`, and instead defines an instance in terms of `DialectPrint`.

In the process we had to define a DialectPrint instance for those dialect that didn't yet have one but did rely on the Repr instance existing. In particular, I've modernized the PaperExamples.lean file to show of the more ergonomic ways of defining a dialect.