feat: cache-mount all subproject .lake folders in Docker build (#1707)
This PR tweaks the Docker build to enable incremental caching (via
cache-mount) of the .lake folder of all sub-projects as well.

In addition, we move the docker caching setup and teardown to a
dedicated shell script, in a new `.dockerscripts` folder, to reduce
caching-related noise in the Dockerfile.

This fixes #1687.