feat/Blase: add bool sort (#1702)
This sort is in mutual with BV, so we convert Term into an indexed
inductive. This changes basically every 'def', but should be mechanical
to change.

At the end of the PR, our internal core IR should be able to represent
`Bool` natively.

A similar treatment can then be performed (much more smoothly, once the
work for `Bool` is done) for `Nat` and `Int`. This PR performs the bulk
of the work needed for the transition, as the bulk of the work is in
going from a single inductive `Term` to an indexed inductive with two
sorts, `Term bv` and `Term bool`.