refactor: get llvm-mca from PATH (#1669)
This PR drops the hard-coded `LLVM_BUILD_DIR` path from the `run_mca.py`
script, so that the llvm-mca binary is found in the PATH instead. I've
also added instructions to the README on how to add a locally-built LLVM
to the path (let me know if the instructions are unclear!).

By getting the binary from the path, we can also use package-manager
provided `llvm-mca`, instead of forcing everybody to locally build LLVM
at some specific path.

---------

Co-authored-by: Luisa Cicolini <48860705+luisacicolini@users.noreply.github.com>
Co-authored-by: Tobias Grosser <tobias@grosser.es>