feat: add `select_constant_cmp` rewrite pattern (#1712)
This PR adds a new rewrite pattern,
[select_constant_cmp](https://github.com/llvm/llvm-project/blob/cae73be8a6c2d39e5a827fd31848676d52062a94/llvm/include/llvm/Target/GlobalISel/Combine.td#L518).

The pattern rewrites cases where select is given constant inputs (_true_
or _false_), simplifying it to the first or second input value
accordingly.

In the current test case, the expected output still contains the input
given to the rewriter. Ideally, this should be eliminated by DCE, which
is not applied in this PR.

Co-authored-by: Luisa Cicolini luisacicolini@gmail.com

---------

Co-authored-by: Alex Keizer <alex@keizer.dev>