feat: expand SLLVM dialect with memory operations (#1719)
This PR expands the SLLVM dialect with basic memory operations (`load`,
`store`, and `ptradd`).

Ideally, we'd have SLLVM re-use the syntax and semantics of `LLVM.Op`
directly from the definition in the `LLVM` dialect, but
we ran into some difficulties with simplification of signatures and
semantics when trying this, so in this PR we've decided to just
duplicate the mapping from operations to signature and semantics (but
obviously re-using the definitions in the LLVM.Semantics file for the
actual semantic content).