chore: remove namespace CI testing (#1731)
This was an experiment to try to pass state from one GH action job to
another. We evaluated three approaches:

- Use namespace artifacts

Just uploading the artifact of `.lake/packages` takes 45s. Downloading
it again takes 8s.

  This is too slow to be useful.

- Use namespace cache

The namespace cache does not reliably propagate to the next jobs, such
that parts of the build is rebuilt. Hence, this is also not a solution.

- GitHub Cache is not perfect but only takes 2 * 8s to download and
works reliably.

GH Action Cache is already used in our standard evaluation flow. Hence,
do changes needed to our core CI. We remove this CI flow to reduce our
CI cost.