chore: add a testcase that exposes a bug in elimValuation simproc (#1746)
This PR adds a minimized example from opencompl/circomvent where the
`elimValuation` simproc generates a wrong proof, thus triggering a
`(kernel) match application` error.

For now, the line that actually triggers the error is commented out, so
that the test doesn't fail and we can merge this PR. When we get around
to fixing this, we should uncomment the line so that the test becomes a
regression test.